+++ packages/util-crypto/src/base32/helpers.ts
@@ -19,6 +19,7 @@ interface Config {
   ipfs?: string;
   regex?: RegExp;
   type: string;
+  withPadding?: boolean;
 }
 
 type DecodeFn = (value: string, ipfsCompat?: boolean) => Uint8Array;
@@ -63,7 +64,7 @@ export function createIs (validate: ValidateFn): ValidateFn {
 }
 
 /** @internal */
-export function createValidate ({ chars, ipfs, type }: Config): ValidateFn {
+export function createValidate ({ chars, ipfs, type, withPadding }: Config): ValidateFn {
   return (value?: unknown, ipfsCompat?: boolean): value is string => {
     if (typeof value !== 'string') {
       throw new Error(`Expected ${type} string input`);
@@ -72,12 +73,17 @@ export function createValidate ({ chars, ipfs, type }: Config): ValidateFn {
     }
 
     for (let i = (ipfsCompat ? 1 : 0), count = value.length; i < count; i++) {
-      if (!(chars.includes(value[i]) || (
-        value[i] === '=' && (
-          (i === value.length - 1) ||
-          !chars.includes(value[i + 1])
-        )
-      ))) {
+      if (chars.includes(value[i])) {
+        // all ok, character found
+      } else if (withPadding && value[i] === '=') {
+        if (i === count - 1) {
+          // last character, everything ok
+        } else if (value[i + 1] === '=') {
+          // next one is also padding, sequence ok
+        } else {
+          throw new Error(`Invalid ${type} padding sequence "${value[i]}${value[i + 1]}" at index ${i}`);
+        }
+      } else {
         throw new Error(`Invalid ${type} character "${value[i]}" (0x${value.charCodeAt(i).toString(16)}) at index ${i}`);
       }
     }